GUI-258. Separate entities from api: - add src/common for types and utils, used in app and worker; - currency move to common/types; - account move to app/api/types; - etherium address tcomb check move to common/types/runtime;
