+++ front/src/app/api/mock.ts
@@ -1,10 +1,7 @@
 import * as t from './types';
-import { ICurrencyInfo, ICurrencyBalanceMap } from 'app/entities/currency';
-import {
-    IAccountInfo,
-    EnumProfileStatus,
-    IProfileBrief,
-} from 'app/entities/account';
+import { ICurrencyInfo } from 'common/types/currency';
+import { EnumProfileStatus, IProfileBrief } from 'app/entities/account';
+import { ICurrencyBalanceMap, IAccountInfo } from 'app/api/types/account';
 
 const VASYA_ADDR = '0x88057f14236687831e1fd205e8efb9e45166fe72';
 const PETYA_ADDR = '0xfd0c80ba15cbf19770319e5e76ae05012314608f';
