+++ front/src/common/types/currency.ts
@@ -1,0 +1,7 @@
+export interface ICurrencyInfo {
+    symbol: string;
+    decimalPointOffset: number;
+    name: string;
+    address: string;
+    balance: string;
+}
